+++ x/emissions/keeper/queryserver/query_server_score_test.go
@@ -8,31 +8,32 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-func (s *QueryServerTestSuite) TestGetLatestInfererScore() {
+func (s *QueryServerTestSuite) TestGetInfererScoreEma() {
 	ctx := s.ctx
 	keeper := s.emissionsKeeper
 	topicId := uint64(1)
 	worker := "worker1"
-	oldScore := types.Score{TopicId: topicId, BlockHeight: 1, Address: worker, Score: alloraMath.NewDecFromInt64(90)}
 	newScore := types.Score{TopicId: topicId, BlockHeight: 2, Address: worker, Score: alloraMath.NewDecFromInt64(95)}
 
 	// Set an initial score for inferer and attempt to update with an older score
-	_ = keeper.SetLatestInfererScore(ctx, topicId, worker, newScore)
-	err := keeper.SetLatestInfererScore(ctx, topicId, worker, oldScore)
-	s.Require().NoError(err, "Setting an older inferer score should not fail but should not update")
+	err := keeper.SetInfererScoreEma(ctx, topicId, worker, newScore)
+	s.Require().NoError(err, "Setting an inferer score should not fail")
 
-	req := &types.QueryLatestInfererScoreRequest{
+	req := &types.QueryGetInfererScoreEmaRequest{
 		TopicId: topicId,
 		Inferer: worker,
 	}
-	response, err := s.queryServer.GetLatestInfererScore(ctx, req)
+	response, err := s.queryServer.GetInfererScoreEma(ctx, req)
 	s.Require().NoError(err)
 
-	updatedScore := response.Score
-	s.Require().NotEqual(oldScore.Score, updatedScore.Score, "Older score should not replace newer score")
+	s.Require().True(
+		newScore.Score.Equal(response.Score.Score),
+		"Score should be set %s | %s",
+		newScore.Score.String(),
+		response.Score.Score.String())
 }
 
-func (s *QueryServerTestSuite) TestGetLatestForecasterScore() {
+func (s *QueryServerTestSuite) TestGetForecasterScoreEma() {
 	ctx := s.ctx
 	keeper := s.emissionsKeeper
 	topicId := uint64(1)
@@ -41,20 +42,20 @@ func (s *QueryServerTestSuite) TestGetLatestForecasterScore() {
 	newScore := types.Score{TopicId: topicId, BlockHeight: 2, Address: worker, Score: alloraMath.NewDecFromInt64(95)}
 
 	// Set a new score for forecaster
-	_ = keeper.SetLatestForecasterScore(ctx, topicId, forecaster, newScore)
+	_ = keeper.SetForecasterScoreEma(ctx, topicId, forecaster, newScore)
 
-	req := &types.QueryLatestForecasterScoreRequest{
+	req := &types.QueryGetForecasterScoreEmaRequest{
 		TopicId:    topicId,
 		Forecaster: forecaster,
 	}
-	response, err := s.queryServer.GetLatestForecasterScore(ctx, req)
+	response, err := s.queryServer.GetForecasterScoreEma(ctx, req)
 	s.Require().NoError(err)
 
 	forecasterScore := response.Score
 	s.Require().Equal(newScore.Score, forecasterScore.Score, "Newer forecaster score should be set")
 }
 
-func (s *QueryServerTestSuite) TestGetLatestReputerScore() {
+func (s *QueryServerTestSuite) TestGetReputerScoreEma() {
 	ctx := s.ctx
 	keeper := s.emissionsKeeper
 	topicId := uint64(1)
@@ -63,13 +64,13 @@ func (s *QueryServerTestSuite) TestGetLatestReputerScore() {
 	newScore := types.Score{TopicId: topicId, BlockHeight: 2, Address: worker, Score: alloraMath.NewDecFromInt64(95)}
 
 	// Set a new score for reputer
-	_ = keeper.SetLatestReputerScore(ctx, topicId, reputer, newScore)
+	_ = keeper.SetReputerScoreEma(ctx, topicId, reputer, newScore)
 
-	req := &types.QueryLatestReputerScoreRequest{
+	req := &types.QueryGetReputerScoreEmaRequest{
 		TopicId: topicId,
 		Reputer: reputer,
 	}
-	response, err := s.queryServer.GetLatestReputerScore(ctx, req)
+	response, err := s.queryServer.GetReputerScoreEma(ctx, req)
 	s.Require().NoError(err)
 
 	reputerScore := response.Score